+++ .docz/app/imports.js
@@ -3,6 +3,10 @@ export const imports = {
     import(
       /* webpackPrefetch: true, webpackChunkName: "src-index" */ 'src/index.mdx'
     ),
+  'src/components/EditForm/EditView.mdx': () =>
+    import(
+      /* webpackPrefetch: true, webpackChunkName: "src-components-edit-form-edit-view" */ 'src/components/EditForm/EditView.mdx'
+    ),
   'src/components/DetailsView/DetailsView.mdx': () =>
     import(
       /* webpackPrefetch: true, webpackChunkName: "src-components-details-view-details-view" */ 'src/components/DetailsView/DetailsView.mdx'
@@ -15,17 +19,13 @@ export const imports = {
     import(
       /* webpackPrefetch: true, webpackChunkName: "src-components-details-view-single-details-view" */ 'src/components/DetailsView/SingleDetailsView.mdx'
     ),
-  'src/components/EditForm/EditView.mdx': () =>
+  'src/components/Labels/Label.mdx': () =>
     import(
-      /* webpackPrefetch: true, webpackChunkName: "src-components-edit-form-edit-view" */ 'src/components/EditForm/EditView.mdx'
+      /* webpackPrefetch: true, webpackChunkName: "src-components-labels-label" */ 'src/components/Labels/Label.mdx'
     ),
   'src/components/Labels/Title.mdx': () =>
     import(
       /* webpackPrefetch: true, webpackChunkName: "src-components-labels-title" */ 'src/components/Labels/Title.mdx'
-    ),
-  'src/components/Labels/Label.mdx': () =>
-    import(
-      /* webpackPrefetch: true, webpackChunkName: "src-components-labels-label" */ 'src/components/Labels/Label.mdx'
     ),
   'src/components/EditForm/EditField/DateTimeField.mdx': () =>
     import(
