+++ src/components/EditForm/EditField/SelectField/index.tsx
@@ -1,6 +1,9 @@
-import { MenuProps, NoticeProps } from 'react-select/lib/components/Menu';
-import React, { CSSProperties, HTMLAttributes, useMemo } from 'react';
-import TextField, { BaseTextFieldProps } from '@material-ui/core/TextField';
+import {
+  FieldWrapperProps,
+  SelectGroupOption,
+  SelectOption
+} from '@src/components/TypeDefinitions';
+import React, { CSSProperties, useMemo } from 'react';
 import {
   Theme,
   createStyles,
@@ -8,69 +11,14 @@ import {
   useTheme
 } from '@material-ui/core/styles';
 
-import CancelIcon from '@material-ui/icons/Cancel';
-import Chip from '@material-ui/core/Chip';
-import { ControlProps } from 'react-select/lib/components/Control';
-import { FieldWrapperProps } from '@src/components/TypeDefinitions';
-import MenuItem from '@material-ui/core/MenuItem';
-import { MultiValueProps } from 'react-select/lib/components/MultiValue';
 import NoSsr from '@material-ui/core/NoSsr';
 import { Omit } from '@material-ui/types';
-import { OptionProps } from 'react-select/lib/components/Option';
-import Paper from '@material-ui/core/Paper';
-import { PlaceholderProps } from 'react-select/lib/components/Placeholder';
 import Select from 'react-select';
-import { SingleValueProps } from 'react-select/lib/components/SingleValue';
-import Typography from '@material-ui/core/Typography';
-import { ValueContainerProps } from 'react-select/lib/components/containers';
+import { StylesConfig } from 'react-select/lib/styles';
 import { ValueType } from 'react-select/lib/types';
-import clsx from 'clsx';
+import components from './CustomComponents';
 import { emphasize } from '@material-ui/core/styles/colorManipulator';
-
-interface OptionType {
-  label: string;
-  value: string;
-}
-
-const suggestions: OptionType[] = [
-  { label: 'Afghanistan' },
-  { label: 'Aland Islands' },
-  { label: 'Albania' },
-  { label: 'Algeria' },
-  { label: 'American Samoa' },
-  { label: 'Andorra' },
-  { label: 'Angola' },
-  { label: 'Anguilla' },
-  { label: 'Antarctica' },
-  { label: 'Antigua and Barbuda' },
-  { label: 'Argentina' },
-  { label: 'Armenia' },
-  { label: 'Aruba' },
-  { label: 'Australia' },
-  { label: 'Austria' },
-  { label: 'Azerbaijan' },
-  { label: 'Bahamas' },
-  { label: 'Bahrain' },
-  { label: 'Bangladesh' },
-  { label: 'Barbados' },
-  { label: 'Belarus' },
-  { label: 'Belgium' },
-  { label: 'Belize' },
-  { label: 'Benin' },
-  { label: 'Bermuda' },
-  { label: 'Bhutan' },
-  { label: 'Bolivia, Plurinational State of' },
-  { label: 'Bonaire, Sint Eustatius and Saba' },
-  { label: 'Bosnia and Herzegovina' },
-  { label: 'Botswana' },
-  { label: 'Bouvet Island' },
-  { label: 'Brazil' },
-  { label: 'British Indian Ocean Territory' },
-  { label: 'Brunei Darussalam' }
-].map(suggestion => ({
-  value: suggestion.label,
-  label: suggestion.label
-}));
+import linq from 'linq';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -139,142 +87,24 @@ const useStyles = makeStyles((theme: Theme) =>
   })
 );
 
-const NoOptionsMessage = ({
-  selectProps,
-  innerProps,
-  children
-}: NoticeProps<OptionType>) => (
-  <Typography
-    color="textSecondary"
-    className={selectProps.classes.noOptionsMessage}
-    {...innerProps}
-  >
-    {children}
-  </Typography>
-);
-
-type InputComponentProps = Pick<BaseTextFieldProps, 'inputRef'> &
-  HTMLAttributes<HTMLDivElement>;
-
-const inputComponent = ({ inputRef, ...props }: InputComponentProps) => (
-  <div ref={inputRef} {...props} />
-);
-
-const Control = ({
-  selectProps,
-  innerRef,
-  children,
-  innerProps
-}: ControlProps<OptionType>) => (
-  <TextField
-    fullWidth
-    InputProps={{
-      inputComponent,
-      inputProps: {
-        className: clsx({
-          [selectProps.classes.input]: true,
-          [selectProps.classes[selectProps.TextFieldProps.variant]]: true
-        }),
-        inputRef: innerRef,
-        children: children,
-        ...innerProps
-      }
-    }}
-    {...selectProps.TextFieldProps}
-  />
-);
-
-const Option = ({
-  innerRef,
-  isFocused,
-  isSelected,
-  innerProps,
-  children
-}: OptionProps<OptionType>) => (
-  <MenuItem
-    ref={innerRef}
-    selected={isFocused}
-    component="div"
-    style={{
-      fontWeight: isSelected ? 500 : 400
-    }}
-    {...innerProps}
-  >
-    {children}
-  </MenuItem>
-);
-
-const Placeholder = ({
-  children,
-  selectProps,
-  innerProps
-}: PlaceholderProps<OptionType>) => (
-  <Typography
-    color="textSecondary"
-    className={selectProps.classes.placeholder}
-    {...innerProps}
-  >
-    {children}
-  </Typography>
-);
-
-const SingleValue = ({
-  children,
-  selectProps,
-  innerProps
-}: SingleValueProps<OptionType>) => (
-  <Typography className={selectProps.classes.singleValue} {...innerProps}>
-    {children}
-  </Typography>
-);
-
-const ValueContainer = ({
-  selectProps,
-  children
-}: ValueContainerProps<OptionType>) => (
-  <div className={selectProps.classes.valueContainer}>{children}</div>
-);
-
-const MultiValue = ({
-  children,
-  selectProps,
-  removeProps,
-  isFocused,
-  options
-}: MultiValueProps<OptionType>) => {
-  const index = options.findIndex(o => o.value === children) % 3;
-  return (
-    <Chip
-      color={index === 0 ? 'primary' : index === 1 ? 'secondary' : 'default'}
-      tabIndex={-1}
-      label={children}
-      className={clsx(selectProps.classes.chip, {
-        [selectProps.classes.chipFocused]: isFocused
-      })}
-      onDelete={removeProps.onClick}
-      deleteIcon={<CancelIcon {...removeProps} />}
-    />
-  );
-};
-
-const Menu = ({ selectProps, innerProps, children }: MenuProps<OptionType>) => (
-  <Paper square className={selectProps.classes.paper} {...innerProps}>
-    {children}
-  </Paper>
-);
-
-const components = {
-  Control,
-  Menu,
-  MultiValue,
-  NoOptionsMessage,
-  Option,
-  Placeholder,
-  SingleValue,
-  ValueContainer
-};
-
 type VariantType = 'standard' | 'filled' | 'outlined';
+
+const getGroupOptions = (
+  options: Array<SelectOption>
+): Array<SelectOption> | Array<SelectGroupOption> => {
+  const query = linq.from(options);
+  const hasGroup = query.any(i => i.group !== undefined);
+  return hasGroup
+    ? query
+        .groupBy(i => i.group || '')
+        .orderBy(g => g.key())
+        .select(g => ({
+          label: g.key(),
+          options: g.toArray()
+        }))
+        .toArray()
+    : options;
+};
 
 interface SelectFieldProps extends Omit<FieldWrapperProps, 'variant'> {
   variant?: VariantType;
@@ -289,18 +119,23 @@ function SelectField({
   variant,
   required,
   multiSelection,
+  options,
   ...rest
 }: SelectFieldProps) {
   const classes = useStyles();
   const theme = useTheme();
-  const [single, setSingle] = React.useState<ValueType<OptionType>>(null);
+  const [single, setSingle] = React.useState<ValueType<SelectOption>>(null);
 
-  function handleChangeSingle(value: ValueType<OptionType>) {
+  function handleChangeSingle(value: ValueType<SelectOption>) {
     setSingle(value);
   }
 
   const selectStyles = useMemo(
-    () => ({
+    (): StylesConfig => ({
+      indicatorsContainer: (base: CSSProperties) => ({
+        ...base,
+        cursor: 'pointer'
+      }),
       input: (base: CSSProperties) => ({
         ...base,
         color: theme.palette.text.primary,
@@ -322,7 +157,7 @@ function SelectField({
         name={name}
         classes={classes}
         styles={selectStyles}
-        options={suggestions}
+        options={getGroupOptions(options || [])}
         components={components}
         value={single}
         isMulti={multiSelection}
@@ -333,6 +168,7 @@ function SelectField({
           label: (variant as any) === 'labeled' ? undefined : label,
           variant: (variant as any) === 'labeled' ? 'outlined' : variant,
           required,
+
           InputLabelProps: {
             shrink: placeholder || single ? true : undefined
           }
